requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Bcrypt
Flask-Cors
python-dotenv
Flask
gunicorn